update workqueue and routeplan

+++ backend/controllers/RouteplanController.php
@@ -38,12 +38,17 @@ class RouteplanController extends Controller
      */
     public function actionIndex()
     {
+        $pageSize = \Yii::$app->request->post("perpage");
+
         $searchModel = new RouteplanSearch();
         $dataProvider = $searchModel->search($this->request->queryParams);
+
+        $dataProvider->pagination->pageSize = $pageSize;
 
         return $this->render('index', [
             'searchModel' => $searchModel,
             'dataProvider' => $dataProvider,
+            'perpage' => $pageSize,
         ]);
     }
 
@@ -70,7 +75,36 @@ class RouteplanController extends Controller
         $model = new Routeplan();
 
         if ($this->request->isPost) {
-            if ($model->load($this->request->post()) && $model->save()) {
+            if ($model->load($this->request->post())) {
+                $drop_off_place = \Yii::$app->request->post('drop_off_place');
+                $drop_off_qty = \Yii::$app->request->post('drop_off_qty');
+
+                if ($model->save(false)) {
+                    if (count($drop_off_place)) {
+                        for ($i = 0; $i <= count($drop_off_place) - 1; $i++) {
+                            $route_plan_line_chk = \common\models\RoutePlanLine::find()->where(['route_plan_id' => $model->id, 'dropoff_place_id' => $drop_off_place[$i]])->one();
+                            if ($route_plan_line_chk) {
+                                $route_plan_line_chk->dropoff_place_id = $drop_off_place[$i];
+                                $route_plan_line_chk->dropoff_qty = $drop_off_qty[$i];
+                                $route_plan_line_chk->status = $model->status;
+                                if ($route_plan_line_chk->save(false)) {
+
+                                }
+                            } else {
+                                $new_line = new \common\models\RoutePlanLine();
+                                $new_line->route_plan_id = $model->id;
+                                $new_line->dropoff_place_id = $drop_off_place[$i];
+                                $new_line->dropoff_qty = $drop_off_qty[$i];
+                                $new_line->status = $model->status;
+                                if ($new_line->save(false)) {
+
+                                }
+                            }
+
+                        }
+                    }
+                }
+
                 return $this->redirect(['view', 'id' => $model->id]);
             }
         } else {
@@ -93,12 +127,50 @@ class RouteplanController extends Controller
     {
         $model = $this->findModel($id);
 
-        if ($this->request->isPost && $model->load($this->request->post()) && $model->save()) {
+        $model_line = \common\models\RoutePlanLine::find()->where(['route_plan_id'=>$model->id])->all();
+
+        if ($this->request->isPost && $model->load($this->request->post())) {
+            $drop_off_place = \Yii::$app->request->post('drop_off_place');
+            $drop_off_qty = \Yii::$app->request->post('drop_off_qty');
+
+            $removelist = \Yii::$app->request->post('remove_list');
+
+            if ($model->save(false)) {
+                if (count($drop_off_place)) {
+                    for ($i = 0; $i <= count($drop_off_place) - 1; $i++) {
+                        $route_plan_line_chk = \common\models\RoutePlanLine::find()->where(['route_plan_id' => $model->id, 'dropoff_place_id' => $drop_off_place[$i]])->one();
+                        if ($route_plan_line_chk) {
+                            $route_plan_line_chk->dropoff_place_id = $drop_off_place[$i];
+                            $route_plan_line_chk->dropoff_qty = $drop_off_qty[$i];
+                            $route_plan_line_chk->status = $model->status;
+                            if ($route_plan_line_chk->save(false)) {
+
+                            }
+                        } else {
+                            $new_line = new \common\models\RoutePlanLine();
+                            $new_line->route_plan_id = $model->id;
+                            $new_line->dropoff_place_id = $drop_off_place[$i];
+                            $new_line->dropoff_qty = $drop_off_qty[$i];
+                            $new_line->status = $model->status;
+                            if ($new_line->save(false)) {
+
+                            }
+                        }
+                    }
+                }
+            }
+
+            $delete_rec = explode(",", $removelist);
+            if (count($delete_rec)) {
+                \common\models\RoutePlanLine::deleteAll(['id' => $delete_rec]);
+            }
+
             return $this->redirect(['view', 'id' => $model->id]);
         }
 
         return $this->render('update', [
             'model' => $model,
+            'model_line'=>$model_line,
         ]);
     }
 
@@ -111,7 +183,16 @@ class RouteplanController extends Controller
      */
     public function actionDelete($id)
     {
-        $this->findModel($id)->delete();
+//        $this->findModel($id)->delete();
+
+        $model_line = \common\models\RoutePlanLine::find()->where(['route_plan_id'=>$id])->all();
+        if ($model_line){
+            if (\common\models\RoutePlanLine::deleteAll(['route_plan_id' => $id])){
+                $this->findModel($id)->delete();
+            }
+        }else{
+            $this->findModel($id)->delete();
+        }
 
         return $this->redirect(['index']);
     }
